add support for both mist repos (mist and mist cli); update descriptions
added 'github_repo' to list of inputs to allow changing which version of mist is downloaded. this input will  become the value for the 'github_repo' key under GitHubReleasesInfoProvider

add inline comment to show acceptable repo's for mist/mist-cli

update descriptions to reflect the two versions; update munki description since it's not downloading mist

adds mist-cli path to install_item_paths which will add a file and the respective checksum if mist-cli is downloaded; will not add the path if the respective package isn't downloaded in autopkg.